backend/requirements.txt
fastapi
python-dotenv
google-generativeai
uvicorn
sqlalchemy 
psycopg2
alembic
asyncpg
loguru
starlette
python-jose
authlib
